+++ tools/lsp/Cargo.toml
@@ -49,7 +49,7 @@ backend-gl-all = ["backend-winit", "renderer-winit-femtovg"]
 backend-gl-wayland = ["backend-winit-wayland", "renderer-winit-femtovg"]
 backend-gl-x11 = ["backend-winit-x11", "renderer-winit-femtovg"]
 
-preview = ["slint-interpreter", "i-slint-backend-selector"]
+preview = ["slint-interpreter", "i-slint-core", "i-slint-backend-selector"]
 
 default = ["backend-qt", "backend-winit", "renderer-winit-femtovg", "preview"]
 
@@ -60,10 +60,10 @@ euclid = "0.22"
 lsp-types = { version = "0.93.0", features = ["proposed"] }
 serde = "1.0.118"
 serde_json = "1.0.60"
-i-slint-core = { version = "=0.3.1", path = "../../internal/core" }
 
 
 # for the preview
+i-slint-core = { version = "=0.3.1", path = "../../internal/core", optional = true }
 slint-interpreter = { version = "=0.3.1", path = "../../internal/interpreter", default-features = false, features = ["compat-0-3-0"], optional = true  }
 i-slint-backend-selector = { version = "=0.3.1", path="../../internal/backends/selector", optional = true }
